Moved Input and ParseResult into types module, made primitives a real module
